exercise 4.21, step 9

+++ part4/controllers/blogs.js
@@ -34,6 +34,21 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
+
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token invalid'})
+  }
+  const blog = await Blog.findByIdAndDelete(request.params.id)
+
+  if (!blog) {
+    return response.status(204).end()
+  }
+
+  if (blog.user.toString() !== decodedToken.id) {
+    return response.status(401).json({ error: 'only the creator can delete blogs' })
+  }
   await Blog.findByIdAndDelete(request.params.id)
   response.status(204).end()
 })
